+++ android-json-form-wizard/src/main/java/com/vijay/jsonwizard/filesource/AssetsFileSource.java
@@ -4,6 +4,7 @@ import android.content.Context;
 
 import com.vijay.jsonwizard.constants.JsonFormConstants;
 import com.vijay.jsonwizard.interfaces.FormFileSource;
+import com.vijay.jsonwizard.utils.FileReaderUtil;
 
 import org.jeasy.rules.api.Rules;
 import org.jeasy.rules.mvel.MVELRuleFactory;
@@ -42,16 +43,7 @@ public class AssetsFileSource implements FormFileSource {
             InputStream inputStream = context.getAssets()
                     .open(newFileName);
 
-            BufferedReader reader = new BufferedReader(new InputStreamReader(inputStream,
-                    "UTF-8"));
-            String jsonString;
-            StringBuilder stringBuilder = new StringBuilder();
-            while ((jsonString = reader.readLine()) != null) {
-                stringBuilder.append(jsonString);
-            }
-            inputStream.close();
-
-            return new JSONObject(stringBuilder.toString());
+            return new JSONObject(FileReaderUtil.convertStreamToString(inputStream));
         } catch (IOException e) {
             Timber.e(e);
         } catch (JSONException e) {
