+++ android-json-form-wizard/src/main/java/com/vijay/jsonwizard/utils/FileReaderUtil.java
@@ -1,12 +1,14 @@
-package org.smartregister.nativeform.util;
+package com.vijay.jsonwizard.utils;
 
 import android.content.Context;
 
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileInputStream;
+import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
+import java.io.UnsupportedEncodingException;
 
 public class FileReaderUtil {
 
@@ -18,8 +20,9 @@ public class FileReaderUtil {
         return ret;
     }
 
-    public static String convertStreamToString(InputStream is) throws Exception {
-        BufferedReader reader = new BufferedReader(new InputStreamReader(is));
+    public static String convertStreamToString(InputStream is) throws IOException {
+        BufferedReader reader = new BufferedReader(new InputStreamReader(is,
+                "UTF-8"));
         StringBuilder sb = new StringBuilder();
         String line;
         while ((line = reader.readLine()) != null) {
