+++ sample/src/main/java/org/smartregister/nativeform/domain/JsonForm.java
@@ -6,10 +6,10 @@ import android.support.annotation.WorkerThread;
 
 import com.vijay.jsonwizard.constants.JsonFormConstants;
 import com.vijay.jsonwizard.domain.Form;
+import com.vijay.jsonwizard.utils.FileReaderUtil;
 
 import org.json.JSONObject;
 import org.smartregister.nativeform.contract.FormTesterContract;
-import org.smartregister.nativeform.util.FileReaderUtil;
 
 import java.io.File;
